requirements.txt
3==0.0.0
alabaster==0.7.12
asgiref==3.4.1
Babel==2.9.1
backports.entry-points-selectable==1.1.1
bcrypt==3.2.2
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.10
crispy-bootstrap5==0.4
cryptography==37.0.4
distlib==0.3.4
Django==3.2.5
django-bootstrap-datetimepicker==1.5
django-crispy-forms==1.12.0
django-leaflet==0.28.1
docutils==0.17.1
Fabric3==1.14.post1
filelock==3.4.0
GDAL==3.0.4
gunicorn==20.1.0
idna==3.3
imagesize==1.3.0
importlib-metadata==4.8.2
importlib-resources==5.4.0
invoke==1.7.1
Jinja2==3.0.3
MarkupSafe==2.0.1
packaging==21.3
paramiko==2.11.0
pathlib2==2.3.7.post1
Pillow==8.3.1
platformdirs==2.4.0
psycopg2==2.9.2
psycopg2-binary==2.9.2
pycparser==2.21
Pygments==2.11.2
PyNaCl==1.5.0
pyparsing==3.0.6
pytz==2021.1
reportlab==3.6.1
requests==2.27.1
six==1.16.0
snowballstemmer==2.2.0
Sphinx==4.3.2
sphinx-rtd-theme==1.0.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.7
virtualenv==20.10.0
zipp==3.6.0
